widget/tests/test_key_event_counts.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<!-- We've had issues on Mac OS X where native key events either don't get processed
     or they get processed twice. This test tests some of those scenarios. -->

<window id="window1" title="Test Key Event Counts" onload="runTest()"
  xmlns:html="http://www.w3.org/1999/xhtml"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/NativeKeyCodes.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <pre id="test"></pre>
  </body>

  <script type="application/javascript"><![CDATA[
    var gKeyPressEventCount = 0;

    function onKeyPress(e)
    {
      gKeyPressEventCount++;
      e.preventDefault();
    }

    function runTest()
    {
      var domWindowUtils = window.QueryInterface(Components.interfaces.nsIInterfaceRequestor).
                             getInterface(Components.interfaces.nsIDOMWindowUtils);

      window.addEventListener("keypress", onKeyPress, false);

      // Test ctrl-tab
      gKeyPressEventCount = 0;
      domWindowUtils.sendNativeKeyEvent(0, MAC_VK_Tab, 0x0400, "\t", "\t");
      is(gKeyPressEventCount, 1);

      // Test cmd+shift+a
      gKeyPressEventCount = 0;
      domWindowUtils.sendNativeKeyEvent(0, MAC_VK_ANSI_A, 0x4000 | 0x0100, "a", "A");
      is(gKeyPressEventCount, 1);

      // Test cmd-;
      gKeyPressEventCount = 0;
      domWindowUtils.sendNativeKeyEvent(0, MAC_VK_ANSI_Semicolon, 0x4000, ";", ";");
      is(gKeyPressEventCount, 1);

      window.removeEventListener("keypress", onKeyPress, false);
    }
  ]]></script>

</window>
